+++ .gitignore
@@ -1,0 +1,73 @@
+# Archivos generados por Python
+__pycache__/
+*.pyc
+*.pyo
+*.pyd
+*.pyo
+*.pyw
+*.pyz
+*.pyzw
+
+# Ambiente virtual
+venv/
+env/
+*.venv/
+
+# Archivos de configuración
+*.pyc
+*.pyo
+*.pyd
+
+# Archivos de logs y resultados de pruebas
+*.log
+*.pot
+*.pyc
+__pycache__/
+htmlcov/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+.hypothesis/
+
+# Dependencias
+*.egg
+*.egg-info/
+dist/
+build/
+eggs/
+parts/
+var/
+sdist/
+develop-eggs/
+.installed.cfg
+
+# Herramientas específicas
+.tox/
+docs/_build/
+docs/api/
+
+# Sublime Text
+*.tmlanguage.cache
+*.tmPreferences.cache
+*.stTheme.cache
+
+# Visual Studio Code
+.vscode/
+
+# Jupyter Notebook
+.ipynb_checkpoints/
+
+# Ignore .DS_Store files on macOS
+.DS_Store
+
+# Ignore Thumbs.db files on Windows
+Thumbs.db
+
+# Ignore .idea folder for PyCharm
+.idea/
+
+# Ignore .vscode folder for Visual Studio Code
+.vscode/
