+++ app.py
@@ -1,18 +1,287 @@
-from flask import Flask
+from flask import Flask, request
 
-from blueprints.curvaturas import curvaturas_bp
-from blueprints.formas_fundamentales import formas_fundamentales_bp
-from blueprints.representacion import representacion_bp
-from blueprints.caracteristicas import caracteristicas_bp
+from utils import *
 
+"""
+-------------------------------------------------------------------------------
+FUNCIONES AUXILIARES
+-------------------------------------------------------------------------------
+"""
 
+def normaliza_parametrizacion(var1, var2, sup):
+    """
+    Transforma la parametrización de una superficie a una expresion sympy con sus variable correspondientes
+    No se hacen comprobaciones de tipo
+
+    Argumentos:
+    var1                string con la primera variable de la parametrización
+    var2                string con la segunda variable de la parametrización
+    sup                 string con la parametrización de la superficie
+    """
+    if var1:
+        u = sp.symbols(var1, real = True)
+    else:
+        u = sp.symbols('u', real = True)
+        var1='u'
+
+    if var2:
+        v = sp.symbols(var2, real = True)
+    else:
+        v = sp.symbols('v', real = True)
+        var2='v'
+
+    elementos_str = sup.strip('[]').split(',')
+    try:
+        superficie = [sp.sympify(elem, locals={var1: u, var2: v}) for elem in elementos_str]
+    except Exception as e:
+        raise Exception(f"Error al procesar la superficie: {e}")
+    if len(superficie) != 3:
+        raise Exception(f"La parametrización de la superficie debe tener 3 elementos pero se han encontrado {len(superficie)}")
+    
+    return superficie, u, v
+
+"""
+-------------------------------------------------------------------------------
+ENDPOINTS
+-------------------------------------------------------------------------------
+"""
 app = Flask(__name__)
 
-# Registrar blueprints
-app.register_blueprint(curvaturas_bp)
-app.register_blueprint(formas_fundamentales_bp)
-app.register_blueprint(representacion_bp)
-app.register_blueprint(caracteristicas_bp)
+@app.route('/primera_forma_fundamental')
+def primera_forma_fundamental():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+    
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(primeraFormaFundamental_pt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(primeraFormaFundamental_pt_xyz(superficie, u, v, x0, y0, z0))
+    
+    return str(primeraFormaFundamental(superficie, u, v))
+
+@app.route('/segunda_forma_fundamental')
+def segunda_forma_fundamental():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+    
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(segundaFormaFundamental_pt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(segundaFormaFundamental_pt_xyz(superficie, u, v, x0, y0, z0))
+    
+    return str(segundaFormaFundamental(superficie, u, v))
+
+@app.route('/curvatura_Gauss')
+def curvatura_Gauss():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+    
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(curvaturaGauss_pt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(curvaturaGauss_pt_xyz(superficie, u, v, x0, y0, z0))
+    
+    return str(curvaturaGauss(superficie, u, v))
+
+@app.route('/curvatura_media')
+def curvatura_media():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(curvaturaMedia_pt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(curvaturaMedia_pt_xyz(superficie, u, v, x0, y0, z0))
+    
+    return str(curvaturaMedia(superficie, u, v))
+
+@app.route('/curvaturas_principales')
+def curvaturas_principales():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(curvaturasPrincipales_pt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(curvaturasPrincipales_pt_xyz(superficie, u, v, x0, y0, z0))
+
+    
+    return str(curvaturasPrincipales(superficie, u, v))
+
+@app.route('/vector_normal')
+def vector_normal():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+    
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(normal_pt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(normal_pt_xyz(superficie, u, v, x0, y0, z0))
+    
+    return str(normal(superficie, u, v))
+
+@app.route('/clasificacion_punto')
+def clasificacion_punto():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+    
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(clasicPt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(clasicPt_xyz(superficie, u, v, x0, y0, z0))
+    
+    raise Exception("No se ha definido correctamente el punto a clasificar")
+
+@app.route('/plano_tangente')
+def plano_tangente():
+    var1 = request.args.get('var1', None)
+    var2 = request.args.get('var2', None)
+    superficie_str  = request.args.get('superficie')
+    
+    if not superficie_str:
+        raise Exception("No se ha encontrado la parametrización de la superficie")
+
+    try:
+        superficie, u, v = normaliza_parametrizacion(var1, var2, superficie_str)
+    except Exception as e:
+        #TODO-que hacer si hay error?
+        raise e
+    
+    u0 = request.args.get('u0', None)
+    v0 = request.args.get('v0', None)
+    if u0 and v0:
+        return str(planoTangente_pt_uv(superficie, u, v, u0, v0))
+
+    x0 = request.args.get('x0', None)
+    y0 = request.args.get('y0', None)
+    z0 = request.args.get('z0', None)
+    if x0 and y0 and z0:
+        return str(planoTangente_pt_xyz(superficie, u, v, x0, y0, z0))
+    
+    return str(planoTangente(superficie, u, v))
+
+@app.route('/direcciones_principales')
+def direcciones_principales():
+    # TODO
+    return ''
+
+@app.route('/representacion')
+def representar():
+    # TODO
+    return ''
+
 
 if __name__ == '__main__':
     app.run(debug=True)
